+++ src/constants.ts
@@ -76,6 +76,8 @@ interface Neuro {
     currentController: string | null;
     /** If this is `true`, all permissions return `Off`. */
     killSwitch: boolean;
+    /** The last known user selection, or null if there is none or it is in a non-Neuro-safe file. */
+    lastKnownUserSelection: vscode.Selection | null;
 }
 
 
@@ -108,6 +110,7 @@ export const NEURO: Neuro = {
     highlightDecorationType: null,
     currentController: null,
     killSwitch: false,
+    lastKnownUserSelection: null,
 };
 
 // this will likely be transformed for a different use later when the API rolls around